requirements.txt
torch>=2.0
wandb
hydra-core
omegaconf
tqdm
Pillow
matplotlib
numpy
